+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.0.70] - 2025-02-13
 
 - Fix in `setSetting` action where the setting was not updated in the workspace configuration
+- Added actions to the Demo Time panel menu
 - [#49](https://github.com/estruyf/vscode-demo-time/issues/#49): New command to create a new demo file
 
 ## [0.0.69] - 2025-02-12